TAG-248 Flytt typer til komponenten den blir brukt
